+++ mod_detector.py
@@ -10,12 +10,12 @@ class ModDetector(tk.Toplevel):
         self.parent = parent
         self.title("Mod Detector")
         self.geometry("550x270")
-        # self.resizable(width=False, height=False)
+        self.resizable(width=False, height=False)
         self.transient(parent)
         self.grab_set()
 
-        self.minecraft_location = os.getenv("APPDATA") + "\\.minecraft"
-        self.minecraft_mods = self.minecraft_location + "\\mods"
+        self.minecraft_location = parent.minecraft_location
+        self.minecraft_mods = parent.minecraft_mods
 
         self.widget_frame_main = ttk.Frame(self)
         self.widget_frame_main.pack(side="top", fill="both")
@@ -30,7 +30,8 @@ class ModDetector(tk.Toplevel):
 
         ttk.Style().configure("All.TButton", font=("", "10", "underline"))
 
-        self.widget_button_right_all = ttk.Button(self.widget_frame_buttons, text=">>>", command=self.move_all_mods_right, style="All.TButton")
+        self.widget_button_right_all = ttk.Button(self.widget_frame_buttons, text=">>>",
+                                                  command=self.move_all_mods_right, style="All.TButton")
         self.widget_button_right_all.grid(row=0, column=0, sticky="we")
 
         self.widget_button_right = ttk.Button(self.widget_frame_buttons, text=">>>", command=self.move_mod_right)
@@ -39,7 +40,8 @@ class ModDetector(tk.Toplevel):
         self.widget_button_left = ttk.Button(self.widget_frame_buttons, text="<<<", command=self.move_mod_left)
         self.widget_button_left.grid(row=2, column=0, sticky="we")
 
-        self.widget_button_left_all = ttk.Button(self.widget_frame_buttons, text="<<<", command=self.move_all_mods_left, style="All.TButton")
+        self.widget_button_left_all = ttk.Button(self.widget_frame_buttons, text="<<<",
+                                                 command=self.move_all_mods_left, style="All.TButton")
         self.widget_button_left_all.grid(row=3, column=0, sticky="we")
 
         self.widget_tree_right = Tree(self.widget_frame_main)
@@ -63,8 +65,7 @@ class ModDetector(tk.Toplevel):
                                                          index="end",
                                                          text=os.path.splitext(file)[0],
                                                          values=(os.path.splitext(file)[1]),
-                                                         tags=os.path.join(self.minecraft_mods, file).replace("\\",
-                                                                                                              "/"))
+                                                         tags=os.path.join(self.minecraft_mods + "/", file))
 
     def move_mod_right(self):
         try:
@@ -142,7 +143,7 @@ def main():
     app.widget_frame_buttons = ttk.Frame(app)
     app.widget_frame_buttons.pack(side="bottom", fill="x")
 
-    app2 = ModDetector(app)
+    ModDetector(app)
     app.mainloop()